Check if path is writable before downloading
Legendary can download many hours with displaying multiple times write errors as demonstrated in https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/issues/3474. This is bad way, it should close immediately or retry maybe only 5 times if file can't be written.
So I'll workaround that by writing temporary file to directory to check if it's writable, if not, then display error and force user to choose another directory.
This will be helpful for cases like #26.